+++ src/Model/Content.php
@@ -42,32 +42,32 @@ class Content
     public function __construct(string $title, string $summaryText = "", string $productDescription = "")
     {
         // validate the title
-        $title = trim($title);
-        if (strlen($title) < self::TITLE_MIN_LENGTH) {
+        $trimmedTitle = trim($title);
+        if (strlen($trimmedTitle) < self::TITLE_MIN_LENGTH) {
             throw new ContentException(sprintf("The title should not be shorter than %s characters",
                 self::TITLE_MIN_LENGTH));
         }
 
-        if (strlen($title) > self::TITLE_MAX_LENGTH) {
+        if (strlen($trimmedTitle) > self::TITLE_MAX_LENGTH) {
             throw new ContentException(sprintf("The title should not be longer than %s characters",
                 self::TITLE_MAX_LENGTH));
         }
 
         // validate the summary
-        $summaryText = trim($summaryText);
+        $trimmedSummary = trim($summaryText);
 
-        if (strlen($summaryText) < self::SUMMARY_MIN_LENGTH) {
+        if (strlen($trimmedSummary) < self::SUMMARY_MIN_LENGTH) {
             throw new ContentException(sprintf("The summary text should not be shorter than %s characters",
                 self::SUMMARY_MIN_LENGTH));
         }
 
-        if (strlen($summaryText) > self::SUMMARY_MAX_LENGTH) {
+        if (strlen($trimmedSummary) > self::SUMMARY_MAX_LENGTH) {
             throw new ContentException(sprintf("The summary text should not be longer than %s characters",
                 self::SUMMARY_MAX_LENGTH));
         }
 
-        $this->title = $title;
-        $this->summaryText = $summaryText;
+        $this->title = $trimmedTitle;
+        $this->summaryText = $trimmedSummary;
         $this->productDescription = $productDescription;
     }