+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+- Exclude the tests/ folder from Scrutinizer reports
+- Fix coding style and security issues reported by Scrutinizer
+
 ## [v0.1.0] - 2016-07-12
 ### Added
 - A composer.json, README and LICENSE