+++ CHANGELOG.md
@@ -4,4 +4,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- A composer.json and a README
+- A composer.json, README and LICENSE
+- Base implementation of wambo/module-catalog with a read-only JSONCatalogProvider