+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+### Changed
+- Renamed the Wambo\Catalog\Error namespace to Wambo\Catalog\Exception
+- Replace the CatalogProvider with a ProductRepository
+- Remove the Catalog Model
+
+### Fixed
+- Exclude the tests/ folder from Scrutinizer reports
+- Fix coding style and security issues reported by Scrutinizer
+
+## [v0.1.0] - 2016-07-12
 ### Added
 - A composer.json, README and LICENSE
 - Base implementation of wambo/module-catalog with a read-only JSONCatalogProvider